requirements.txt
flask
opencv-python
face_recognition
numpy
psycopg2-binary
boto3